+++ controllers/action_router.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, HTTPException, Depends
 import traceback
 
-from domain.action_service import create_actions_for_race, create_action_for_quartered_event
+from domain.action_service import create_actions_for_elimination_event, create_actions_for_race, create_action_for_quartered_event
 from domain.dtos.blob_competitor_dto import BlobCompetitorDto
 from domain.event_service import get_event_by_id
 from domain.exceptions.event_not_found_exception import EventNotFoundException
@@ -40,6 +40,28 @@ def race(event_id: int, tick: int, _=Depends(require_auth)):
 
     try:
         create_actions_for_race(event.competitors, event_id, tick)
+    except Exception as e:
+        traceback.print_exc()
+        raise HTTPException(status_code=500, detail=f"{e.with_traceback(None)}")
+
+
+@router.post("/create/elimination")
+def elimination(contenders: list[BlobCompetitorDto], event_id: int, _=Depends(require_auth)):
+    """
+    Generate score for all contenders and save the actions for given event.
+    """
+    try:
+        _ = get_event_by_id(event_id, check_date=True)
+    except NoCurrentEventException:
+        raise HTTPException(status_code=400, detail="EVENT_IS_CONCLUDED")
+    except EventNotFoundException:
+        raise HTTPException(status_code=404, detail="EVENT_NOT_FOUND")
+
+    try:
+        record = create_actions_for_elimination_event(contenders, event_id)
+        if record is not None:
+            return {"name": record[0], "score": record[1]}
+        return None
     except Exception as e:
         traceback.print_exc()
         raise HTTPException(status_code=500, detail=f"{e.with_traceback(None)}")